+++ src/gui/status_bar.cpp
@@ -61,8 +61,8 @@ void StatusBar::lookAndFeelChanged()
 
 void StatusBar::paintOverChildren(juce::Graphics &g)
 {
-    auto const bounds = this->getLocalBounds().toFloat();
-    auto const thickness = 1.f;
+    auto const bounds = this->getLocalBounds();
+    auto const thickness = 1;
 
     g.setColour(this->findColour(ColorID::ForegroundLow));
     g.fillRect(bounds.getX(), bounds.getY(), bounds.getWidth(), thickness);
